% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\docType{data}
\name{cities}
\alias{cities}
\title{Latitudes and Longitudes of a few locations in Victoria, Australia}
\format{A data frame with 5 rows and three columns
\describe{
  \item{city}{name of city}
  \item{longitude}{longitude in decimal degrees}
  \item{latitude}{latitude in decimal degrees}
}}
\usage{
cities
}
\description{
cities.
}
\keyword{datasets}
