% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maorify.R
\name{maorify}
\alias{maorify}
\title{Substitute regular vowels for Maori language equivalents with macrons}
\usage{
maorify(x)
}
\arguments{
\item{x}{a string}
}
\value{
a string with unicode macroned vowels substituted for "@vowel" substrings.
}
\description{
The Maori language includes vowels with macrons.
Including these in R plots often involves typing unicode strings for the relevant macron vowel
This function makes this easier, by substituting all instances of the substring "@vowel" for the macronised equivalent in unicode
}
\examples{
hist(rpois(1000, 5), xlab=maorify("Number of T@u@i per hectare"),
	 main=maorify("Density of T@u@i at Whakat@ane"), col="blue")
}
\author{
Michael Scroggie
}
