% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toVicgrid.R
\name{toEPSG}
\alias{toEPSG}
\alias{toVicgrid}
\alias{toMGA55}
\alias{toMGA54}
\alias{toWGS84}
\title{Transform Spatial* objects among coordinate systems I commonly use}
\usage{
toEPSG(spobj, EPSG = 3111)
}
\arguments{
\item{spobj}{A Spatial* object.}

\item{EPSG}{An EPSG code}
}
\value{
A Spatial* object in the specified coordinate system.
}
\description{
\code{toEPSG} converts a Spatial* object to another coordinate system, based on EPSG code
Actual conversion is done by \code{spTransform} from package \code{maptools}
}
\details{
This should work for any EPSG code provided at \url{http://spatialreference.org/ref/epsg/}

\code{toVicgrid(spobj)} is a convenient shortcut to convert to Vicgrid (epsg 3111)

\code{toMGA55(spobj)} is a convenient shortcut to convert to MGA55

\code{toMGA54(spobj)} is a convenient shortcut to convert to MGA54

\code{toWGS84(spobj)} is a convenient shortcut to convert to WGS84
}
\examples{
cities_wgs<-SpatialPointsDataFrame(coords=cities[,-1], data=data.frame(city=cities$city), proj4string=CRS("+init=epsg:4326"))
citiesVG<-toEPSG(cities_wgs, 3111)
plot(citiesVG, xlab="Easting", ylab="Northing", main="Victorian cities in Vicgrid coordinate system")
axis(1)
axis(2)
box()
}
\author{
Michael Scroggie
}
