% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toVicgrid.R
\name{toVicgrid}
\alias{toVicgrid}
\title{Transform Spatial* objects among coordinate systems I commonly use}
\usage{
toVicgrid(spobj)
}
\arguments{
\item{spobj}{A Spatial* object.}
}
\value{
A Spatial* object in the specified coordinate system.
}
\description{
Transform Spatial* objects among coordinate systems I commonly use
}
\examples{
cities_wgs<-SpatialPointsDataFrame(coords=cities[,-1], data=data.frame(city=cities$city), proj4string=CRS("+init=epsg:4326"))
citiesVG<-toVicgrid(cities_wgs)
plot(citiesVG, xlab="Easting", ylab="Northing", main="Victorian cities in Vicgrid coordinate system")
axis(1)
axis(2)
box()
}
